Automator workflow to quit System Preferences
Pre-built workflow saved as an Application.  

Quits the System Preferences application if it is running.  This is left open when the Restart to Bootcamp workflow completes (because the system reboots before you can close the app).  Adding this to your Login Items just makes the world a cleaner place.

You can open it with Automator to verify what it does (and doesn't) do.